# Accelerated  memcpy