@digetx 2018-11-20
@kusma commits are updated, please take a look. If it looks okay now, then I can merge this all sometime later after some more testing. I made some extra changes..

1) now copying won't be broken in 4K chunks and will be a single-large "VFP" copy when possible
2) decided to keep memcpy() for cases of copying from cacheable memory because I got a bit different result this time during testing and memcpy could be ~20% faster for specific cases of cacheline aliasing
3) changed the rule of when to use bounce buffer, it will be used only when both src and dest are uncached, perf tests showed that it is more optimal
4) src address is now always aligned, only dst can be bounced.. perf tests showed that it is more optimal