@digetx 2018-11-22
@kusma  Patches updated.

Dropped "offset aligning" copying, we can return to it later once there will be a real use-case (accelerated DownloadFromScreen for example). 

Tuned copying performance, there is now tegra_copy_block_vfp_arm() which is a mix of VFP + STMIA, it should be faster than generic memcpy() in all cases of copying sizes.. etc.

 BLOCK_SIZE reduced from 4K to 1K, this gives some boost for 2-pass copying.

vfpcpy() now doesn't clobber D8-D15 VFP register and copies 64B at a time since I can't see perf difference in comparison to 128B.