@digetx 2018-11-24
Added `__thread` annotation to `bounce_buf`. It is not really necessary for a single-threaded Xorg, but now those vfp-copy functions should be thread-safe.