@digetx 2018-12-01
@kusma Any more comments? I've been testing (using) opentegra with these patches included for a week without any problems.